+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed
+
+- Copy custom game addons from mods to new instances
+
 ## [1.1.6] 2026-08-28
 
 ### Changed